+++ Native.Csharp.Tool/Properties/AssemblyInfo.cs
@@ -32,5 +32,5 @@ using System.Runtime.InteropServices;
 // 可以指定所有值，也可以使用以下所示的 "*" 预置版本号和修订号
 //通过使用 "*"，如下所示:
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion ("3.0.0.0504")]
-[assembly: AssemblyFileVersion ("3.0.0.0504")]
+[assembly: AssemblyVersion ("3.0.1.0510")]
+[assembly: AssemblyFileVersion ("3.0.1.0510")]
