+++ Native.Csharp/FodyWeavers.xml
@@ -1,14 +1,4 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Weavers xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="FodyWeavers.xsd">
-	<Costura>
-		<!--在这里填写不被 Fody 打包的程序集名称, 一行一个--><!--
-		<IncludeAssemblies>
-			
-		</IncludeAssemblies>
-		
-		--><!--在这里填写以 bin\x86\Debug(Release) 为基础的非托管dll的相对路径--><!--
-		<Unmanaged32Assemblies>
-			
-		</Unmanaged32Assemblies>-->
-	</Costura>
+﻿<Weavers xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="FodyWeavers.xsd">
+  <Costura />
+  <ModuleInit />
 </Weavers>
