+++ Native.Csharp/FodyWeavers.xsd
@@ -90,6 +90,7 @@
             </xs:attribute>
           </xs:complexType>
         </xs:element>
+        <xs:element name="ModuleInit" minOccurs="0" maxOccurs="1" type="xs:anyType" />
       </xs:all>
       <xs:attribute name="VerifyAssembly" type="xs:boolean">
         <xs:annotation>
