+++ Native.Csharp/Repair/Core/Kernel32.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using Native.Csharp.Repair.Enum;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Runtime.InteropServices;
 using System.Text;
 
-namespace Native.Csharp.Repair
+namespace Native.Csharp.Repair.Core
 {
 	/*
 	 *	移植自: 00.00.dotnetRedirect 插件, 原作者: 成音S. 引用请带上此注释
