+++ Native.Csharp/Repair/Enum/LoadLibraryFlags.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace Native.Csharp.Repair
+namespace Native.Csharp.Repair.Enum
 {
 	/*
 	 *	移植自: 00.00.dotnetRedirect 插件, 原作者: 成音S. 引用请带上此注释
