+++ Native.Csharp/Repair/Helper/ReflectionHelper.cs
@@ -14,26 +14,39 @@ namespace Native.Csharp.Repair
 	public static class ReflectionHelper
 	{
 		#region --字段--
-		public static BindingFlags bindFlags = BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static; 
+		public static BindingFlags bindFlags = BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic | BindingFlags.Static;
 		#endregion
 
 		#region --公开方法--
 		public static T InvokeMethod<T> (Type type, object instance, string methodName, params object[] args)
 		{
 			var method = GetMethod (type, methodName);
-			return (T)method?.Invoke (instance, args);
+			if (method != null)
+			{
+				return (T)method.Invoke (instance, args);
+			}
+			return default (T);
 		}
 
 		public static T InvokeMethod<T> (this T obj, string methodName, params object[] args)
 		{
 			var type = typeof (T);
 			var method = GetMethod (type, methodName);
-			return (T)method?.Invoke (obj, args);
+			if (method != null)
+			{
+				return (T)method.Invoke (obj, args);
+			}
+			return default (T);
 		}
 
 		public static MethodInfo GetMethod (Type type, string methodName)
 		{
-			return type.GetMethods ().Where (mi => mi.Name == methodName).FirstOrDefault () ?? null;
+			var result = type.GetMethods ().Where (mi => mi.Name == methodName).FirstOrDefault ();
+			if (result != null)
+			{
+				return result;
+			}
+			return null;
 		}
 
 		public static T GetInstanceField<T> (Type type, object instance, string fieldName)
@@ -51,8 +64,10 @@ namespace Native.Csharp.Repair
 		public static void ClearEventInvocations (this object obj, string eventName)
 		{
 			var fi = obj.GetType ().GetEventField (eventName);
-			if (fi == null) return;
-			fi.SetValue (obj, null);
+			if (fi != null)
+			{
+				fi.SetValue (obj, null);
+			}
 		}
 
 		public static FieldInfo GetEventField (this Type type, string eventName)
@@ -62,15 +77,19 @@ namespace Native.Csharp.Repair
 			{
 				field = type.GetField (eventName, bindFlags);
 				if (field != null && (field.FieldType == typeof (MulticastDelegate) || field.FieldType.IsSubclassOf (typeof (MulticastDelegate))))
+				{
 					break;
+				}
 
 				field = type.GetField ("EVENT_" + eventName.ToUpper (), bindFlags);
 				if (field != null)
+				{
 					break;
+				}
 				type = type.BaseType;
 			}
 			return field;
-		} 
+		}
 		#endregion
 	}
 }
