+++ Native.Csharp/packages.config
@@ -1,6 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
-  <package id="Costura.Fody" version="3.2.1" targetFramework="net40" developmentDependency="true" />
-  <package id="Fody" version="3.3.3" targetFramework="net40" developmentDependency="true" />
+  <package id="Costura.Fody" version="4.0.0" targetFramework="net40" />
+  <package id="Fody" version="5.0.5" targetFramework="net40" developmentDependency="true" />
+  <package id="ModuleInit.Fody" version="2.0.0" targetFramework="net40" />
   <package id="Unity" version="5.8.6" targetFramework="net40" />
 </packages>
