+++ README.md
@@ -43,6 +43,16 @@
 > 4. ~~对于接收消息时, 颜文字表情, 特殊符号乱码, 当前正在寻找转换方式~~ <font color=#FF0000>(已修复)</font>
 
 ## Native.SDK 更新日志
+> 2019年05月10日 版本: V3.0.1.0510
+
+	说明: 由于酷Q改动了应用加载机制, 将所有开发中的应用都迁移至 dev 文件夹下, 所以本次更新将针对此优化进行改动
+
+    1. 修复 Native.Csharp.Repair 应用兼容组件运行时间不正确
+    2. 新增 SDK 生成前自动检查是否存在 AppID 目录
+    3. 新增 SDK 生成前自动清理旧版 app.dll app.json 文件
+    4. 新增 SDK 生成后自动复制并重命名程序集为 app.dll 自动复制并重命名json文件为 app.json
+    5. 新增 SDK 生成后自动清理原生成目录下的程序集和 json 文件
+
 > 2019年05月04日 版本: V3.0.0.0504
 
 	说明: 由于酷Q改动了应用机制, 因此升级时请务必保存代码, 进行代码迁移升级!
